Move and fix some examples
