+++ .pipelex/inference/routing_profiles.toml
@@ -29,7 +29,8 @@ default = "pipelex_inference"
 [profiles.pipelex_first.routes]
 # Pattern matching: "model-pattern" = "backend-name"
 "gpt-*" = "pipelex_inference"
-"claude-*" = "pipelex_inference"
+# "claude-*" = "pipelex_inference"
+"claude-*" = "bedrock"
 "grok-*" = "pipelex_inference"
 "gemini-*" = "pipelex_inference"
 "*-sdxl" = "fal"
