+++ examples/extract_dpe/extract_dpe_struct.py
