+++ examples/wip/discord_newsletter/discord_newsletter.plx
