+++ examples/wip/discord_newsletter/discord_newsletter_struct.py
